+++ cmake/Modules/FindLZ4_EP.cmake
@@ -39,6 +39,7 @@ find_library(LZ4_LIBRARIES
   NAMES
     liblz4${CMAKE_STATIC_LIBRARY_SUFFIX}
     liblz4_static${CMAKE_STATIC_LIBRARY_SUFFIX}
+    lz4${CMAKE_STATIC_LIBRARY_SUFFIX}
   PATHS ${TILEDB_EP_INSTALL_PREFIX}
   PATH_SUFFIXES lib
   NO_DEFAULT_PATH
@@ -77,52 +78,28 @@ if (NOT LZ4_FOUND)
   if (TILEDB_SUPERBUILD)
     message(STATUS "Adding LZ4 as an external project")
     if (WIN32)
-      ExternalProject_Add(ep_lz4
-        PREFIX "externals"
-        URL "https://github.com/lz4/lz4/releases/download/v1.8.2/lz4_v1_8_2_win64.zip"
-        URL_HASH SHA1=330f97a88b91a4b9a15ddfe063321849a3a9e62d
-        CONFIGURE_COMMAND ""
-        BUILD_IN_SOURCE TRUE
-        BUILD_COMMAND ""
-        INSTALL_COMMAND
-          ${CMAKE_COMMAND} -E make_directory ${TILEDB_EP_INSTALL_PREFIX}/bin &&
-          ${CMAKE_COMMAND} -E make_directory ${TILEDB_EP_INSTALL_PREFIX}/lib &&
-          ${CMAKE_COMMAND} -E make_directory ${TILEDB_EP_INSTALL_PREFIX}/include &&
-          ${CMAKE_COMMAND} -E copy_if_different
-            ${TILEDB_EP_BASE}/src/ep_lz4/dll/liblz4.def
-            ${TILEDB_EP_BASE}/src/ep_lz4/dll/liblz4.lib
-            ${TILEDB_EP_INSTALL_PREFIX}/bin &&
-          ${CMAKE_COMMAND} -E copy_if_different
-            ${TILEDB_EP_BASE}/src/ep_lz4/dll/liblz4.so.1.8.2.dll
-            ${TILEDB_EP_INSTALL_PREFIX}/bin/liblz4.dll &&
-          ${CMAKE_COMMAND} -E copy_if_different
-            ${TILEDB_EP_BASE}/src/ep_lz4/static/liblz4_static.lib
-            ${TILEDB_EP_INSTALL_PREFIX}/lib &&
-          ${CMAKE_COMMAND} -E copy_if_different
-            ${TILEDB_EP_BASE}/src/ep_lz4/include/lz4.h
-            ${TILEDB_EP_BASE}/src/ep_lz4/include/lz4frame.h
-            ${TILEDB_EP_BASE}/src/ep_lz4/include/lz4hc.h
-            ${TILEDB_EP_INSTALL_PREFIX}/include
-        UPDATE_COMMAND ""
-        LOG_DOWNLOAD TRUE
-        LOG_CONFIGURE TRUE
-        LOG_BUILD TRUE
-      )
-    else()
-      ExternalProject_Add(ep_lz4
-        PREFIX "externals"
-        URL "https://github.com/lz4/lz4/archive/v1.8.2.zip"
-        URL_HASH SHA1=ebf6c227965318ecd73820ade8f5dbd83d48b3e8
-        CONFIGURE_COMMAND ""
-        BUILD_IN_SOURCE TRUE
-        INSTALL_COMMAND $(MAKE) PREFIX=${TILEDB_EP_INSTALL_PREFIX} install
-        UPDATE_COMMAND ""
-        LOG_DOWNLOAD TRUE
-        LOG_CONFIGURE TRUE
-        LOG_BUILD TRUE
-        LOG_INSTALL TRUE
-      )
+      set(ARCH_SPEC -A X64)
     endif()
+    set(LZ4_CMAKE_DIR "${TILEDB_EP_SOURCE_DIR}/ep_lz4/contrib/cmake_unofficial")
+    ExternalProject_Add(ep_lz4
+      PREFIX "externals"
+      URL "https://github.com/lz4/lz4/archive/v1.8.2.zip"
+      URL_HASH SHA1=ebf6c227965318ecd73820ade8f5dbd83d48b3e8
+      CONFIGURE_COMMAND
+        ${CMAKE_COMMAND}
+          ${ARCH_SPEC}
+          -DLZ4_BUILD_LEGACY_LZ4C=OFF
+          -DLZ4_POSITION_INDEPENDENT_LIB=ON
+          -DBUILD_SHARED_LIBS=OFF
+          -DBUILD_STATIC_LIBS=ON
+          -DCMAKE_INSTALL_PREFIX=${TILEDB_EP_INSTALL_PREFIX}
+          ${LZ4_CMAKE_DIR}
+      UPDATE_COMMAND ""
+      LOG_DOWNLOAD TRUE
+      LOG_CONFIGURE TRUE
+      LOG_BUILD TRUE
+      LOG_INSTALL TRUE
+    )
     list(APPEND TILEDB_EXTERNAL_PROJECTS ep_lz4)
   else()
     message(FATAL_ERROR "Unable to find LZ4")